shift scheduler to control-plane
