+++ server/api.go
@@ -7,6 +7,7 @@ import (
 	"github.com/galaxy-io/filament"
 	"github.com/galaxy-io/filament/api/ingestion/v1/ingestionv1connect"
 	"github.com/galaxy-io/filament/eventbus"
+	"github.com/galaxy-io/filament/internal/compile"
 )
 
 type runSubmitter interface {
@@ -23,6 +24,7 @@ type Server struct {
 	bus       eventbus.Bus
 	secrets   filament.Secrets
 	schedules filament.PipelineScheduleStore
+	compiler  *compile.Compiler
 }
 
 // Option configures a Server.
@@ -34,11 +36,12 @@ func WithSecrets(secrets filament.Secrets) Option { return func(s *Server) { s.s
 // New returns a Server wired to the given providers.
 func New(sources filament.SourceRegistry, sinks filament.SinkRegistry, store filament.DataStore, orch runSubmitter, bus eventbus.Bus, opts ...Option) *Server {
 	s := &Server{
-		sources: sources,
-		sinks:   sinks,
-		store:   store,
-		orch:    orch,
-		bus:     bus,
+		sources:  sources,
+		sinks:    sinks,
+		store:    store,
+		orch:     orch,
+		bus:      bus,
+		compiler: &compile.Compiler{Store: store, Sources: sources, Sinks: sinks, DefaultTenant: defaultTenant("")},
 	}
 	if schedules, ok := store.(filament.PipelineScheduleStore); ok {
 		s.schedules = schedules
